requirements.txt
pyserial==3.5
python-dotenv==0.20.0
redis==3.5.3
